poker-vision-object-detector/pyproject.toml
[project]
name = "poker-vision-object-detector"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.uv]
package = false
